separated nlohmann_json from digital_curling::digital_curling
