+++ CMakeLists.txt
@@ -15,7 +15,8 @@ set(CMAKE_CXX_VISIBLIRY_PRESET hidden)
 set(CMAKE_VISIBLITY_INLINES_HIDDEN TRUE)
 
 # main source
-add_subdirectory(digital_curling)
+add_subdirectory(digital_curling) # digital_curling::digital_curling
+add_subdirectory(json)   # digital_curling::json
 
 # apps
 option(DIGITAL_CURLING_BUILD_SERVER "Build digital curling server" ON)
