+++ digital_curling/include/digital_curling/detail/simulation/simulator_fcv1_setting.hpp
@@ -2,7 +2,6 @@
 #define DIGITAL_CURLING_SIMULATION_SIMULATOR_FCV1_SETTING_HPP
 
 #include <memory>
-#include "../json.hpp"
 #include "simulator_setting.hpp"
 
 namespace digital_curling::simulation {
@@ -14,8 +13,6 @@ namespace digital_curling::simulation {
 /// </summary>
 class SimulatorFCV1Setting : public ISimulatorSetting {
 public:
-    static constexpr char kType[] = "fcv1";
-
     /// <summary>
     /// フレームレート(フレーム毎秒)
     /// </summary>
@@ -25,21 +22,9 @@ public:
     virtual ~SimulatorFCV1Setting() = default;
 
     virtual std::unique_ptr<ISimulator> CreateSimulator() const override;
-    virtual void ToJson(nlohmann::json & json) const override;
 };
 
 
-
-} // namespace digital_curling::simulation
-
-
-
-// json
-
-namespace digital_curling::simulation {
-
-void to_json(nlohmann::json &, SimulatorFCV1Setting const&);
-void from_json(nlohmann::json const&, SimulatorFCV1Setting &);
 
 } // namespace digital_curling::simulation
