+++ digital_curling/include/digital_curling/detail/simulation/simulator_setting.hpp
@@ -2,7 +2,6 @@
 #define DIGITAL_CURLING_SIMULATION_SIMULATOR_SETTING_HPP
 
 #include <memory>
-#include "../json.hpp"
 
 namespace digital_curling::simulation {
 
@@ -22,37 +21,10 @@ public:
     /// </summary>
     /// <returns>新たに作成されたシミュレータ</returns>
     virtual std::unique_ptr<ISimulator> CreateSimulator() const = 0;
-
-    /// <summary>
-    /// 実装用関数．JSONを出力する
-    /// </summary>
-    /// <param name="json">出力用JSON</param>
-    virtual void ToJson(nlohmann::json & json) const = 0;
 };
 
 
 
 } // namespace digital_curling::simulation
-
-
-
-// json
-
-namespace digital_curling::simulation {
-
-void to_json(nlohmann::json &, ISimulatorSetting const&);
-
-} // namespace digital_curling::simulation
-
-namespace nlohmann {
-
-template <>
-struct adl_serializer<std::unique_ptr<digital_curling::simulation::ISimulatorSetting>> {
-    static std::unique_ptr<digital_curling::simulation::ISimulatorSetting> from_json(json const&);
-    static void to_json(json &, std::unique_ptr<digital_curling::simulation::ISimulatorSetting> const&);
-};
-
-} // namespace nlohmann
-
 
 #endif // DIGITAL_CURLING_SIMULATION_SIMULATOR_SETTING_HPP
