+++ tests/test/util.hpp
@@ -3,5 +3,6 @@
 
 #include "gtest/gtest.h"
 #include "digital_curling/digital_curling.hpp"
+#include "digital_curling/json.hpp"
 
 #endif // DIGITAL_CURLING_TEST_UTIL_HPP
